services/handwriting-ai/pyproject.toml
[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "handwriting-ai"
version = "0.1.0"
description = "Handwriting AI service for MNIST digit inference (HTTP)."
authors = ["Austin Wagner <austinwagner@msn.com>"]
packages = [
  { include = "handwriting_ai", from = "src" },
  # Include runtime worker scripts as a package for console entry import
  { include = "scripts" },
]
readme = "README.md"
include = ["src/handwriting_ai/py.typed"]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.121.1"
hypercorn = "^0.17.3"
Pillow = "^10.4.0"
numpy = ">=1.26,<3"
python-multipart = "^0.0.20"
typing-extensions = "^4.12.2"
# Torch CPU only (explicit CPU wheels)
torch = {version = "==2.4.1+cpu", source = "pytorch-cpu", markers = "platform_system != 'Darwin' or platform_machine != 'arm64'"}
torchvision = {version = "==0.19.1+cpu", source = "pytorch-cpu", markers = "platform_system != 'Darwin' or platform_machine != 'arm64'"}
redis = "^5.2.0"
rq = "^2.0.0"
httpx = "^0.28.1"
threadpoolctl = "^3.5.0"
psutil = "^7.1.3"
platform-core = { path = "../../libs/platform_core", develop = true }
platform-workers = { path = "../../libs/platform_workers", develop = true }
platform-discord = { path = "../../libs/platform_discord", develop = true }
platform-ml = { path = "../../libs/platform_ml", develop = true }

[tool.poetry.scripts]
handwriting-rq-worker = "handwriting_ai.worker_entry:main"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.6.1"
mypy = "^1.13.0"
ruff = "^0.14.4"
black = "^25.11.0"
yamllint = "^1.37.1"
types-psutil = "^7.0.0.20251111"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "scripts"]
exclude = [".venv"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # bugbear
  "BLE",# blind except
  "UP", # pyupgrade
  "N",  # pep8-naming
  "C4", # flake8-comprehensions
  "SIM",# flake8-simplify
  "RET",# flake8-return
  "C90",# mccabe complexity
  "RUF",# ruff-specific
]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.Any" = { msg = "Do not use typing.Any; prefer precise types or Protocols/TypedDicts." }
"typing.cast" = { msg = "Do not use typing.cast; prefer adapters or precise types." }

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
no_implicit_optional = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
pretty = true
files = ["src", "tests", "scripts"]
# Include local typings and the monorepo guards library
mypy_path = ["src"]
explicit_package_bases = true

[tool.coverage.run]
source = ["src", "scripts"]
omit = ["tests/*"]
branch = true
concurrency = ["multiprocessing", "thread"]

[tool.coverage.report]
precision = 2
show_missing = true
fail_under = 100

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-v -n auto --dist loadscope"
